feat: add dynamic global messages to trip details
